refactor: Update `mainRoutes.js` to `router.js` (#2)
* refactor: Add CI configuration for automated testing

* refactor: Move workflows to the folder `.github/workflows`

* refactor: Update mainRoutes.js to router.js
